Reorg types
