+++ src/types/ProfileTypes.ts
@@ -1,24 +1,5 @@
-import { ActivePersonMissionsOnDay } from "../arena/types";
-
-export type color =
-  | "inherit"
-  | "primary"
-  | "secondary"
-  | "success"
-  | "error"
-  | "info"
-  | "warning"
-  | undefined;
-
-export type buttonType = "button" | "reset" | "submit" | undefined;
-
-export type buttonVariant = "text" | "contained" | "outlined" | undefined;
-
-export type PlanningField = {
-  name: string;
-  code: string;
-  content: string;
-};
+import { PlanningField } from "./ArenaTypes";
+import { ActivePersonMissionsOnDay } from "./ArenaTypes";
 
 export interface Project {
   id: string;
@@ -27,40 +8,6 @@ export interface Project {
   team: Array<any>;
   tools: Array<any>;
 }
-
-export type ActiveExperience = {
-  title: string;
-  esTitle?: string;
-  amount: number;
-  overview: [];
-  esOverview?: string;
-  completed: boolean;
-  priority: string;
-  _type: string;
-  github: string;
-  athleteNotes: string;
-};
-
-export type ActiveMission = {
-  title: string;
-  description: string;
-  proofLink: string;
-};
-
-export type MongoExperience = {
-  [index: string] : any;
-  achievements: number;
-  approved: boolean;
-  description: string;
-  id: string;
-  memberId: string;
-  title: string;
-  type: string;
-  xp: number;
-  xpEarned: number;
-  _id: string;
-  __v: number;
-};
 
 export interface User {
   rank?: number;
@@ -146,52 +93,4 @@ export interface User extends MinimalUser {
 
   createdAt: string;
   __v: number;
-}
-
-export interface Relationship {
-  mentee: User;
-  mentor: User;
-  tasks: Array<any>;
-  resources: Array<any>;
-  sprints: Array<any>;
-  events: Array<any>;
-  reminders: Array<any>;
-  _id: string;
-  id: string;
-  coachId: string;
-  athleteId: string;
-  accepted: boolean;
-  completed: boolean;
-  createdAt: string;
-  __v: number;
-}
-
-export interface Sprint {
-  id: string;
-  athleteId: string;
-  coachId: string;
-  startDate: Date;
-  endDate: Date;
-  started: boolean;
-  events: Array<object>;
-  studySessions: Array<object>;
-  type: string;
-  verified: boolean;
-  teams: Array<User>;
-  stake: {
-    title: string;
-    value: string;
-    conditions: Array<any>;
-  };
-  createdAt: Date;
-  updatedAt: Date;
-}
-
-export type Coach = {
-  _id: string;
-  mentor: {
-    picture: string;
-    fName: string;
-    lName: string;
-  }
 }
