+++ tsconfig.json
@@ -22,8 +22,7 @@
       "@types/testing-library__jest-dom",
       "@types/react"
     ],
-    "typeRoots": ["src/types/custom_types", "node_modules/@types"]
-    // "typeRoots": ["./src/types/custom_types", "./node_modules/@types"]
+    "typeRoots": ["node_modules/@types"]
   },
   "include": ["src", "tests", "jest.setup.js"]
 }
